home view now uses restul api to fetch events
