+++ partEZ/app/Http/Controllers/HomeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\ApiControllers\Views\ApiHomeController;
 use App\Invite;
 use Auth;
 use DB;
@@ -10,7 +11,7 @@ use App\Event;
 use App\Http\Requests;
 use Illuminate\Support\Facades\Request;
 
-class HomeController extends Controller
+class HomeController extends ApiHomeController
 {
     /**
      * Create a new controller instance.
@@ -27,51 +28,8 @@ class HomeController extends Controller
      */
     public function index($local_events_only = false)
     {
-        $user = Auth::user();
-        $events = $this->getUsersEvents();
-        $invites = $this->getUserInvitedEvents();
-        $public_events = Event::getPublicEvents();
-        $this->updateCity();
-
-        return view('home')->with('events', $events)
-            ->with('invites', $invites)
-            ->with('public_events', $public_events)
-            ->with('local_events_only', $local_events_only)
-            ->with('city', $user->city);
+        $response =  ApiHomeController::index();
+        return view('home')->with('data', $response->getData());
     }
 
-    public function updateCity()
-    {
-        $input = Request::all();
-        $user = Auth::user();
-        $city =  array_key_exists ('city' , $input) ? $input['city'] : $user->city;
-        $user->city = $city;
-        $user->save();
-    }
-
-    public function getUsersEvents()
-    {
-        $user = Auth::user();
-
-        $events = Event::getUserEvents($user->uid);
-
-        return $events;
-
-    }
-
-    public function getUserInvitedEvents()
-    {
-        $invites = Invite::getActiveUserInvites();
-        $events = [];
-        foreach($invites as $invite)
-        {
-            $event_array= Event::getByInviteEID($invite->eid);
-            foreach($event_array as $single_event)
-            {
-                array_push($events, $single_event);
-            }
-        }
-
-        return $events;
-    }
 }
